Fix jsx-wrap-multiline rule for closing parenthesis
There was a bug in this rule: having the opening parenthesis on a new
line before the opening tag, but the closing parenthesis on the same
line as the closing tag, would not trigger this rule.
This was caused by the fact that the `index` variable (as calculated by
`siblings.indexOf`) was equal to -1 in some cases.
The `index` is now calculated by checking just the `pos` and `end`
fields of a `node`, instead of the default object equality that
`indexOf` uses.

With this change, the check that tested whether the node started with a
newline is no longer needed (and with it, also the `prevTokenKind`
variable is no longer needed).

Also, two tests have been added:

- A multiline JSX component with the closing parenthesis on a new line,
  but the opening parenthisis on the same line
- A multiline JSX component with the opening parenthesis on a new line,
  but the closing parenthisis on the same line